+++ src/streams.h
@@ -14,7 +14,7 @@ namespace Streams{
 
         void skip(size_t start_sample, size_t end_sample);
 
-        [[nodiscard]] size_t get_duration_s() const;
+        [[nodiscard]] float get_duration_s() const;
 
         [[nodiscard]] size_t get_size() const;
